refactor(client): Decompose RightPanel into custom hook and smaller UI components
